added the possibility to add/edit/delete items

+++ lib/models/item.dart
@@ -1,7 +1,7 @@
 class Item {
   late int id;
   late String name;
-  late int price;
+  late double price;
 
   Item(this.id, this.name, this.price);
 }
